Cross compile .NET and .NET Framework
This allows the package downloader tool to use NuGet.PackageManagement and the website to use NuGet.Versioning and NuGet.Frameworks. This is necessary because the latest NuGet.Commands (a dependency of NuGet.PackageManagement) is not compatible with ASP.NET MVC 2.x (which itself is needed to run on .NET Framework). This is one small step towards moving the project away from .NET Framework entirely. But that's a lot of work.